feat: modfication de fichier et ajout du dossier hooks
